[Refactor] Remove old /Elements code
