+++ src/Apps/Artist/Routes/RelatedArtists/RelatedArtistsList.tsx
@@ -3,7 +3,7 @@ import { RelatedArtistsList_artist } from "__generated__/RelatedArtistsList_arti
 import { ContextConsumer } from "Artsy/Router"
 import React, { Component } from "react"
 import { createRefetchContainer, graphql, RelayRefetchProp } from "react-relay"
-import { Col, Row } from "Styleguide/Elements"
+import { Col, Row } from "Styleguide/Elements/Grid"
 import { Responsive } from "Utils/Responsive"
 
 import {
