+++ src/Apps/Artwork/ArtworkApp.tsx
@@ -6,7 +6,7 @@ import React from "react"
 import { createFragmentContainer, graphql } from "react-relay"
 import styled from "styled-components"
 import { FullArtworkGrid, Tab, Tabs } from "Styleguide/Components"
-import { Col, Row } from "Styleguide/Elements"
+import { Col, Row } from "Styleguide/Elements/Grid"
 import { ArtistInfo } from "./Components/ArtistInfo"
 import { ArtworkDetailsFragmentContainer as ArtworkDetails } from "./Components/ArtworkDetails"
 import { ArtworkSidebarFragmentContainer as ArtworkSidebar } from "./Components/ArtworkSidebar"
