--- src/Styleguide/Elements/__stories__/Arrow.story.tsx
@@ -1,17 +1,0 @@
-import React from "react"
-import { storiesOf } from "storybook/storiesOf"
-import { Section } from "Styleguide/Utils/Section"
-import { Arrow } from "../Arrow"
-
-storiesOf("Styleguide/Elements", module).add("Arrow", () => {
-  return (
-    <React.Fragment>
-      <Section title="Arrow">
-        <Arrow direction="left" />
-        <Arrow direction="right" />
-        <Arrow direction="up" />
-        <Arrow direction="down" />
-      </Section>
-    </React.Fragment>
-  )
-})
