--- src/Styleguide/Elements/__stories__/Flex.story.tsx
@@ -1,15 +1,0 @@
-import { Flex } from "@artsy/palette"
-import React from "react"
-import { storiesOf } from "storybook/storiesOf"
-import { Section } from "Styleguide/Utils/Section"
-
-storiesOf("Styleguide/Elements", module).add("Flex", () => {
-  return (
-    <Section title="Flex">
-      <Flex>
-        A Flex component is a layout container for content that exposes props
-        related flexbox.
-      </Flex>
-    </Section>
-  )
-})
