+++ src/Styleguide/Elements/index.tsx
@@ -1,16 +1,3 @@
 export * from "./Arrow"
-export * from "./Avatar"
-export * from "./Box"
-export * from "./Button"
-export * from "./Checkbox"
-export * from "./Flex"
 export * from "./Grid"
-export * from "./Image"
-export * from "./Join"
 export * from "./Logo"
-export * from "./Message"
-export * from "./Radio"
-export * from "./RadioGroup"
-export * from "./Select"
-export * from "./Separator"
-export * from "./Spacer"
